--- src/hardware/sensors/camera_files/SN33580719.conf
@@ -1,104 +1,0 @@
-[LEFT_CAM_2K]
-fx=1061.39
-fy=1061.41
-cx=1092.88
-cy=626.92
-k1=-0.0455277
-k2=0.0148223
-k3=-0.00632336
-p1=1.28364e-05
-p2=0.000297191
-
-[RIGHT_CAM_2K]
-fx=1060.43
-fy=1060.05
-cx=1102.81
-cy=640.793
-k1=-0.0481335
-k2=0.0178072
-k3=-0.00721597
-p1=9.04589e-05
-p2=-0.000125007
-
-[LEFT_CAM_FHD]
-fx=1061.39
-fy=1061.41
-cx=948.88
-cy=545.92
-k1=-0.0455277
-k2=0.0148223
-k3=-0.00632336
-p1=1.28364e-05
-p2=0.000297191
-
-[RIGHT_CAM_FHD]
-fx=1060.43
-fy=1060.05
-cx=958.81
-cy=559.793
-k1=-0.0481335
-k2=0.0178072
-k3=-0.00721597
-p1=9.04589e-05
-p2=-0.000125007
-
-[LEFT_CAM_HD]
-fx=530.695
-fy=530.705
-cx=632.94
-cy=361.46
-k1=-0.0455277
-k2=0.0148223
-k3=-0.00632336
-p1=1.28364e-05
-p2=0.000297191
-
-[RIGHT_CAM_HD]
-fx=530.215
-fy=530.025
-cx=637.905
-cy=368.3965
-k1=-0.0481335
-k2=0.0178072
-k3=-0.00721597
-p1=9.04589e-05
-p2=-0.000125007
-
-[LEFT_CAM_VGA]
-fx=265.3475
-fy=265.3525
-cx=331.97
-cy=188.23
-k1=-0.0455277
-k2=0.0148223
-k3=-0.00632336
-p1=1.28364e-05
-p2=0.000297191
-
-[RIGHT_CAM_VGA]
-fx=265.1075
-fy=265.0125
-cx=334.4525
-cy=191.69825
-k1=-0.0481335
-k2=0.0178072
-k3=-0.00721597
-p1=9.04589e-05
-p2=-0.000125007
-
-[STEREO]
-Baseline=119.857
-TY=-0.114866
-TZ=-0.290713
-CV_2K=-0.000915054
-CV_FHD=-0.000915054
-CV_HD=-0.000915054
-CV_VGA=-0.000915054
-RX_2K=0.00049945
-RX_FHD=0.00049945
-RX_HD=0.00049945
-RX_VGA=0.00049945
-RZ_2K=0.0017269
-RZ_FHD=0.0017269
-RZ_HD=0.0017269
-RZ_VGA=0.0017269
