--- src/hardware/sensors/config/VLP32C-Onboard.yaml
@@ -1,5 +1,0 @@
-rpm: 600
-returns: Strongest
-fov:
-  start: 0
-  end: 359
