--- src/hardware/sensors/package.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-    <name>sensors</name>
-    <version>0.0.0</version>
-    <description>Driver launchers for car sensors</description>
-    <maintainer email="team@qutmotorsport.com">Alistair English, Alastair Bradford</maintainer>
-    <license>MIT</license>
-
-    <buildtool_depend>ament_cmake</buildtool_depend>
-
-    <exec_depend>ros2launch</exec_depend>
-
-    <export>
-        <build_type>ament_cmake</build_type>
-    </export>
-</package>
