--- src/perception/lidar_pipeline/config/laserscan_params.yaml
@@ -1,14 +1,0 @@
-pointcloud_to_laserscan_node:
-  ros__parameters:
-    target_frame: velodyne
-    transform_tolerance: 0.01
-    min_height: -1.3
-    max_height: -0.5
-    angle_min: -1.5708
-    angle_max: 1.5708
-    angle_increment: 0.004363  # M_PI/720 degrees
-    scan_time: 0.1
-    range_min: 1.0
-    range_max: 25.0
-    use_inf: true
-    inf_epsilon: 1.0
