+++ src/perception/perception_bringup/config/VLP32C-Onboard.yaml
@@ -1,0 +1,5 @@
+rpm: 600
+returns: Strongest
+fov:
+  start: 0
+  end: 359
