+++ src/perception/perception_bringup/config/VLP32C.yaml
@@ -1,0 +1,29 @@
+velodyne_driver_node:
+    ros__parameters:
+        device_ip: 192.168.1.201
+        gps_time: false
+        time_offset: 0.0
+        enabled: true
+        read_once: false
+        read_fast: false
+        repeat_delay: 0.0
+        frame_id: velodyne
+        model: 32C
+        rpm: 600.0
+        port: 2368
+
+velodyne_laserscan_node:
+    ros__parameters:
+        ring: -1
+        resolution: 0.01
+
+velodyne_transform_node:
+    ros__parameters:
+        calibration: VeloView-VLP32C.yaml
+        min_range: 0.5
+        max_range: 130.0
+        view_direction: 0.0
+        view_width: 3.14159
+        organize_cloud: true
+        fixed_frame: ""
+        target_frame: ""
