--- src/perception/vision_pipeline/package.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-    <name>vision_pipeline</name>
-    <version>0.0.0</version>
-    <description>Pipeline for managing visual detection of cones.</description>
-    <maintainer email="team@qutmotorsport.com">Alistair English, Alastair Bradford</maintainer>
-    <license>MIT</license>
-
-    <depend>cv_bridge</depend>
-    <depend>message_filters</depend>
-    <!-- Messages -->
-    <depend>geometry_msgs</depend>
-    <depend>driverless_msgs</depend>
-    <depend>driverless_common</depend>
-    <depend>sensor_msgs</depend>
-    <depend>std_msgs</depend>
-
-    <export>
-        <build_type>ament_python</build_type>
-    </export>
-</package>
