--- src/perception/vision_pipeline/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/vision_pipeline
-[install]
-install_scripts=$base/lib/vision_pipeline
